Make phoenix package expose class with static methods
